requirements.txt
playsound
pystray
Pillow
